requirements.txt
os
sys
requests
hashlib
time
python-dotenv
pandas
numpy
matplotlib
re
nltk
tensorflow
transformers